fix: types minor fixed

+++ src/algorithms/dqn_learning.py
@@ -208,8 +208,8 @@ class DqnLearning:
                 self.write_env_metrics_train()
 
                 if self.episode > int(0.1*self.episodes) and int(np.mean(self.greedy_steps[-int(0.05*self.episodes):])) == self.max_steps:
-                    self.greedy_cumulative_reward = [self.greedy_cumulative_reward[-1]]*(self.episodes/run_greedy_frequency)
-                    self.greedy_steps = [self.greedy_steps[-1]]*(self.episodes/run_greedy_frequency)
+                    self.greedy_cumulative_reward = [self.greedy_cumulative_reward[-1]]*int(self.episodes/run_greedy_frequency)
+                    self.greedy_steps = [self.greedy_steps[-1]]*int(self.episodes/run_greedy_frequency)
                     break
 
                 if (self.episode % run_greedy_frequency) == 0:
