+++ src/algorithms/tlr_learning.py
@@ -175,8 +175,8 @@ class TensorLowRankLearning:
                     self.run_greedy_episode()
 
                 if episode > int(0.1*self.episodes) and int(np.mean(self.greedy_steps[-int(0.05*self.episodes):])) == self.max_steps:
-                    self.greedy_cumulative_reward = [self.greedy_cumulative_reward[-1]]*(self.episodes/run_greedy_frequency)
-                    self.greedy_steps = [self.greedy_steps[-1]]*(self.episodes/run_greedy_frequency)
+                    self.greedy_cumulative_reward = [self.greedy_cumulative_reward[-1]]*int(self.episodes/run_greedy_frequency)
+                    self.greedy_steps = [self.greedy_steps[-1]]*int(self.episodes/run_greedy_frequency)
                     break
         else:
             for _ in range(self.episodes):
